fix(components): Fixes correctedMark function, adds custom compare function & fixes percentage const & more
There's the PercentageProps change which fixes a few typing issues. This led me to fix the calculateCorrectedMark function by creating a custom compare function. Furthermore, I fixed the setPercentages const to ensure it saves the ID & percentage, allowing for the validation of the development project's ID in the function, ensuring it doesn't remove that module's percentage if it happens to be the lowest.

TL;DR - Makes a lot of crucial fixes that makes Year 3's component functional.