feat(routes): Adds Degree Classification & Mark table
This does the above. Firstly we have a useEffect to make sure it runs the getDegreeMark function only if conditions are met. I also fixed a few issues where it was showing Year 2's corrected mark before even selecting anything, and displays the degree mark/classification as a table, making it somewhat more accessible and nicer to look at.